requirements.txt
# Python version
python>=3.8

# Core data science libraries
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# Machine learning libraries
scikit-learn>=1.3.0
joblib>=1.3.0
torch>=2.0.0
xgboost>=1.7.0

# Technical analysis libraries
ta>=0.10.0
pandas-ta>=0.3.14b0
# Указываем конкретную версию ta для совместимости
ta-lib>=0.4.0

# Trading API libraries
ccxt>=4.0.0
tinkoff-investments>=0.2.0b114

# Web framework
flask>=2.3.0
flask-cors>=4.0.0

# Security and environment
python-dotenv>=1.0.0
cryptography>=41.0.0

# Additional utilities
seaborn>=0.12.0